Bug 676602 — Port to GMenu
Split the menu up into an application and a window menu, and port them both
to GMenu/GAction.

The new menu layout is:
 • Application:
  - New Game
  - About Hitori
  - Help
  - Quit
 • Window:
  - Board Size (with submenu)
  - Hint
  - Undo
  - Redo

This was chosen because the actions in the window menu all perform some
operation specifically on the contents of the main window. It’s arguable
that ‘New Game’ should also be there, but that could be considered to open
a new main window, sort of, so it’s in the application menu.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=676602