Make board rendering colours independent of the widget state
The Hitori game board widget doesn’t make use of the GTK+ widget state
(e.g. insensitive, focused) correctly, so the rendering colours should
not be based on it.

Hard-code the colours so that Hitori rendering is independent of the
GTK+ state, until we make the game play area a proper GtkWidget and set
its state flags properly.

https://bugzilla.gnome.org/show_bug.cgi?id=731307