data: Prevent the window being resized
This is a temporary fix for the fact that (for some reason) the window
doesn’t maximise properly. A better fix can be implemented once the
problem’s been investigated (and probably once the GtkDrawingArea has
been replaced with a proper GtkWidget for drawing the board, which
handles resizing correctly).

https://bugzilla.gnome.org/show_bug.cgi?id=736499