build: Use AX_IS_RELEASE
Check for the presence of a git repository to determine whether things
like compiler warnings should be fatal. The warnings themselves will be
hooked up later.

This adds an implicit dependency on autoconf-archive or GNOME’s
m4-common. These are pulled in by JHBuild, and the relevant macros are
automatically distributed in tarballs — so the dependencies only need to
be explicitly installed if you are building from git without using
JHBuild. This is an unsupported use case.

Note that AC_CONFIG_MACRO_DIR() has already been added to configure.ac.

https://wiki.gnome.org/Projects/GnomeCommon/Migration