appdata: Fix project license
The `<project_license>` element is intended to state the license of the
project as a whole. The SPDX `and` term means that the overall license
is the more restrictive union of both given licenses. That means that
the SPDX expression `GPL-3.0+ and CC-BY-SA-3.0` is not FOSS.

This was causing Hitori to be shown as proprietary software in
gnome-software.

See https://github.com/ximion/appstream/issues/312

Fix that by only giving the license of the code as the
`<project_license>`, ignoring the license of the help manual.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>