Incorrect/missing URLs in app metadata
An initiative is under way to to update app metadata URLs, so that they link to up to date web pages and don't link to the wiki, which is going to be retired. The following issues are present for this app:

* MetaInfo `<url type="translate">` should point the app page in https://l10n.gnome.org/

See https://gitlab.gnome.org/GNOME/Initiatives/-/issues/51 for background and guidance.